Add UIKit DiffableListsOfState example
A new `DiffableListsOfState` UIKit example was created to
demonstrate the usage of TCA with a list of (collection view) cells
that have a store and are backed by a diffable data source.

It uses a `Store.child` helper to return a single scoped store
for a particular child (cell), from its identifier. This is useful to
avoid re-scoping all children when only a subset of children needs to
be created (e.g. in a diffing set up, only insertion stores need to be
created).

## Changes

- Create new `DiffableListsOfState` UIKit example
  + Create new DiffableCounterList TCA "stack" and
`DiffableCountersTableViewController` which uses Diffing in a
`UICollectionView` via `UICollectionViewDiffableDataSource`.
  + Create new `CounterCollectionViewCell` which wraps a counter, to
demonstrate injecting child stores in cells.
  + Enrich `CounterState` with a `id: UUID` property for diffing.