Allow multiple strategies in `Effect.throttle`
The `Effect.throttle` operator allowed only two strategies (first,
latest) via the `latest` parameter. While this covers most frequent
scenarios, it doesn't allow any other strategy which is necessary on
some use cases.

One such use case is when we need to throttle values but are interested
in *all* intermediate values, because they can come from different
sources and we can't afford to discard any and rely on just the first
or latest. One concrete example is an e-commerce app where we can
update multiple item quantities' via UI which are then merged and
mapped into rate limited (throttled) network requests to an API, that
contain multiple updates each.

To allow this and other scenarios, a new `Effect.ThrottleStrategy` was
created that's essentially a witness type around a closure responsible
for determining the final value to return in each throttle interval
(essentially a working as a `reduce`).

Backwards compatibility was preserved by keeping the old API, which
forwards to the new one.

## Changes

- Add new `Effect.ThrottleStrategy` witness type.

- Update `Effect.throttle` to receive a new `ThrottleStrategy`.