Add support for task cancellation when using ViewStore.send(_:while).
This adds the ability for sending a cancellation action to the view store when calling
`ViewStore.send(_:while)` inside a SwiftUI .task modifier, in turn making it possible
to start and cancel long-running effects.

Custom logic can be performed by calling ViewStore.send and .suspend directly
with a custom cancellation handler.