[REL-261] ViewStore initializer for void state (#14)
A `ViewStore` that is created from a `Store` with `State == Void` gains little to nothing by subscribing to updates on 
the parent `Store`'s state. Further, in our fork that adds the `removeDuplicates` checks we defaulted to the `==` 
function which will always return `true` for `Void` values, meaning that the updates never made it to the `ViewStore`'s
state and so subscribers (if there were any) would never be updated either.

Within `browser-swift` we currently use the `ViewStore(store.stateless)` pattern exclusively for sending actions back 
into parent stores. This is great and the expected usage but currently we pay for these view stores on the scoping pass
when they are never going to update and have no subscribers. By removing their subscription to their parent store they 
will not take part in anything other than passing sent actions.