[REL-261] WithViewStore special casing for Void state
We already had an override for the case `where State == Void` but it did
not funnel through to the new `ViewState<Void, Action>` initializer
because we were providing a `removeDuplicates` parameter. To make use of
the new `ViewStore` intializer we can change the designated initializer
to take in the `ViewStore` instead of initializing it directly, and then
for the `State == Void` case we pass in the desired `ViewStore` that
bypasses the deduplication check.