[PLIN-201] Inject instrumentation on root store and add method to change callbacks during execution (#24)
The instrumentation has proven to be extremely beneficial to our team
when attempting to address performance issues. However, it kind of
spread like a weed through the API surface of TCA when it could have
simply been injected into the root Store and been fine. The original
reason for not being injected to the root Store was that the system did
not really support updating the callbacks after the root Store was
created.

After chatting with others it seems like there is little chance of any
inconsistencies in data being introduced by updating the instrumentation
callbacks after the root Store is initialized. The reason is that since
the Store is fully main thread affined it should be safe to update the
callbacks on the main thread. By doing this, generally using
`DispatchQueue.main.async` to be outside the TCA system, there is no
chance that a reference to the Instrumentation will be accessed during
the callback changes and thus there should be no data races.