in order to make async code testable, forward a `TBCTestStoreTask`
The concurrency update provides more insight into the runtime of
actions sent into a TestStore.
We didn't forward them to the TBCTestStore interface so far,
but given we're converting more and more things over, this is
now required.

This picks an easy way out, and wraps the `ViewStore`/`TestStore`-Tasks
into a `TBCTestStore` task. Given we're likely moving away from
a custom implementation, this should provide a nice upgrade path.

Todays workshop will also provide some insight if this is enough already
for now, of if we need more.