Avoid failure when setting test dependency values
It's currently possible to emit XCTest failures when testing reducers
that access dependency values that emit failure upon access, which
includes the calendar, locale, time zone, and URL session dependencies.

We may need a more holistic solution for folks that define these kinds
of dependencies outside the library, but this will make the
library-level failures a bit more lenient.