Relax constraints around bindable view state
With #1790 a `BindableStateProtocol` was introduced for deriving
bindings to view state, and some new APIs required this constraint. On
further use, this constraint seems overly strict, as leaf nodes that
don't use view state don't really need this conformance.

While we're at it, let's rename `BindableStateProtocol` to
`BindableViewState`, since it is currently only used to derive view
bindings. If we decide the protocol can do more in the future, we could
consider renaming it to `BindableState` (once the existing property
wrapper type alias is fully removed).