Fix ordered set comparison check (#2082)
For some reason, using `memcmp` with an empty ordered set and a
non-empty ordered set returns true:

```swift
areOrderedSetsDuplicates([1, 2, 3], [])  // true
```

Because of this, we should always check the count before delegating to
`memcmp`. We should also be on the lookout for other exceptions in case
`memcmp` is not appropriate to use here.