PoC requiresOnDeviceRecognition for Speech Recognition Example
Due to how the Speech engine handles emitting SpeechRecognition results, they need to be handled slightly more delicately than just setting the transcription to the resultant string.

I think it would be useful to see in this example some further usage of The Composable Architecture  to show how to dynamically switch the setting at runtime depending on NetworkConditions / etc or perhaps allow the user to dynamically enable and disable their preference for punctuation.