Warn when sending actions from cancelled effects
Today it is possible to cancel a running effect that is sending actions
back into the system and, if the effect does not participate in
cooperative cancellation, be none the wiser.

This PR adds a runtime warning whenever this happens. In general it
shouldn't emit, as cooperatively-cancelling code will have already
thrown an error, but in those cases in which you're not in cooperative
code, it's nice to let the developer know.