Tear down presentation state on dismissal
When a store is initialized with presentation state that is already
presented, we use a flag on `@PresentationState` to determine if we need
to integrate the `\.dismiss` effect to allow the child to dismiss
itself.

This flag currently stays `true` for the lifetime of the presenting
feature, which means if a parent re-presents a feature after dismissal,
the `\.dismiss` effect will fail to do anything when called from the
child.

This PR ensures that the flag is reset to `false` when a presenter
detects that a child feature has been dismissed, fixing this bug.