Annotate the meta-Task returned by Store.send with MainActor
ThreadSanitizer identified a race condition in the handling of the 'tasks'
box in Store.send. It is being accessed by the main thread and by a
background thread (while processing a task). Forcing that
meta task returned by Store.send to run on the Main Actor solves the
race condition warning and seems like the right thing to do.