Fix SwiftUI warning
I noticed this when taking SyncUps for a spin. When I was editing an
attendee and added a new one while the current attendee autocompleted,
I'd get a warning in the `ViewStore`'s `sink` where `objectWillChange`
was getting pinged during a render.

I'm pretty sure we should be dropping the first value from the `sink` to
prevent `objectWillChange` from being pinged when the `ViewStore` is
being initialized in the view.