Better document explicit `id` with scoped `ForEach` (#2784)
`ForEachStore` implicitly used `id: \.state.id` under the hood, so a
naive migration to `ForEach` might miss this detail. Let's be more
explicit in our documentation and migration guide to avoid this issue.

In the future it might be possible to address better in the library
itself, but for now the requirement for stores to be identifiable for
other forms of navigation means allowing for this.