Surface Shared persistence strategies
This branch introduces a second generic to `Shared` that tracks the
underlying persistence strategy:

```swift
Shared<Value, Persistence>
```

This means that a particular persistence strategy can expose endpoints
that the `Shared` could statically call. For example, a remotely-backed
`Shared` could expose a `refresh` endpoint for fetching the latest data:

```swift
@Shared(.api(.config)) var config = Config()

// ...

$config.persistence.refresh()
```

Now this change comes with some complications and source
incompatibilities with earlier periods in the beta. In particular,
`Shared` properties that are _not_ backed by a persistence key can no
longer be specified simply:

```swift
@Shared var count: Int
```

Due to limitations in Swift (lack of default generics, for example), the
new generic cannot be inferred in this situation. The type used to
describe the persistence strategy of such a shared value is `Any`, and
so there is a type alias to simplify:

```swift
// Instead of:
@Shared<Int, Any> var count: Int

// Do:
@AnyShared var count: Int
```

In the future, if Swift ever does get default generics, we could maybe
get rid of these type aliases.

We'd love it if folks could take this change for a spin and/or provide
feedback. We're trying to strike the right balance for shared state in
terms of features and power without sacrificing too much in the way of
usability and API surface area.