Exhaustive shared testing
Just a proof-of-concept on how we could exhaustively test shared state
that is mutated _outside_ of a reducer's state. For example, in SyncUps
we allow the detail screen to remove itself from parent shared state,
but because that state lives outside of the detail screen, it isn't
caught exhaustively. This commit instead checks for unasserted, shared
state when the store deinits, forcing you to exhaustively assert against
it (if the store's exhaustivity is on).

There is more work to do and failing tests to fix. The change tracker
needs to retain changes and not automatically clear them out between
test store assertions, so failures are a bit noisier at the moment, but
assuming we can selectively clear out shared state that lives in the
feature, instead, that should hopefully get things in an ergonomic
place.