Clean up SyncUps tutorial with mock value
We have one step in the tutorial that isn't quite right because it uses
`SyncUp.mock` for a preview when it is never defined.

I think the mock value is probably good to add early in the tutorial as
the same value is used in a bunch of previews, so this commit does just
that!