Avoid unsafe extraction in `BindingReducer`
We currently have a single case of `AnyCasePath(unsafe:)` in the code
base, which uses reflection for extraction, and is something we would
like to remove in a future major version of Case Paths.

This explores a refactoring of `BindingReducer` that takes an explicit
case key path toe the binding action case, instead:

```swift
BindingReducer(action: \.binding)
```

`BindableAction` as a protocol is still necessary for bindable stores to
know how to embed binding actions in a short syntax, _e.g._
`$store.text`.

It's a slight bummer that the reducer gets slightly more verbose,
especially when there _is_ something abstractable here, I'm just not
sure Swift provides the tools yet to truly abstract over a case path in
an ergonomic fashion.

The alternative would be to introduce a special protocol
(`BindableCasePath`?) that the `@Reducer` macro would need to apply
correctly when detecting a `BindableAction`, but this is one-off logic
that seems to take things in the wrong direction.